requirements.txt
python-telegram-bot
bs4
selenium
pydantic
gensim
